requirements.txt
pandas
dash
dash-bootstrap-components
matplotlib
websockets
